Work log 8, 2:30 hours
Spent 1:30 hour bug fixing the score part of my code so that it counted
all of the boxes for each player. Before fixing the score would add
boxes from previous turns and I was able to solve by resetting the
score to zero at the end of the while statement. Since the score is not
displayed until the end, the user does not notice it. I also added in a
segment where the turn repeats if nothing is entered. I also spent 1:00
minutes trying to get my win condition working, but it would always
evaluate as game end on the first turn, so I am trying to add another
turn counter to deal with this problem.